STORE-1012 Allow deleting of entry type

+++ server/openstorefront/openstorefront-core/service/src/main/java/edu/usu/sdl/openstorefront/service/component/CoreComponentServiceImpl.java
@@ -33,6 +33,7 @@ import edu.usu.sdl.openstorefront.core.api.query.QueryType;
 import edu.usu.sdl.openstorefront.core.api.query.SpecialOperatorModel;
 import edu.usu.sdl.openstorefront.core.entity.AlertType;
 import edu.usu.sdl.openstorefront.core.entity.ApprovalStatus;
+import edu.usu.sdl.openstorefront.core.entity.AttributeType;
 import edu.usu.sdl.openstorefront.core.entity.BaseComponent;
 import edu.usu.sdl.openstorefront.core.entity.Component;
 import edu.usu.sdl.openstorefront.core.entity.ComponentAttribute;
@@ -130,6 +131,7 @@ import java.util.Objects;
 import java.util.Set;
 import java.util.concurrent.locks.ReentrantLock;
 import java.util.logging.Level;
+import java.util.logging.Logger;
 import java.util.stream.Collectors;
 import net.java.truevfs.access.TFile;
 import net.java.truevfs.access.TFileInputStream;
@@ -152,6 +154,7 @@ import org.jsoup.select.Elements;
 public class CoreComponentServiceImpl
 		extends BaseComponentServiceImpl
 {
+	protected static final Logger LOG = Logger.getLogger(CoreComponentServiceImpl.class.getName());
 
 	private static final String COPY_MARKER = "- COPY";
 
@@ -608,24 +611,26 @@ public class CoreComponentServiceImpl
 						// Replace converted url
 						String replaceUrl = "LoadMedia&mediaId=".concat(componentMedia.getComponentMediaId());
 						String newUrl = url.substring(0, url.indexOf("TemporaryMedia")).concat(replaceUrl);
-						log.log(Level.FINE, MessageFormat.format("TemporaryMedia Conversion: Replacing {0} with {1}", url, newUrl));
+						super.LOG.log(Level.FINE, MessageFormat.format("TemporaryMedia Conversion: Replacing {0} with {1}", url, newUrl));
 						mediaItem.attr("src", newUrl);
 					} else {
-						log.log(Level.WARNING, MessageFormat.format("Unable to find existing temporary media for temporaryID: {0}", tempMediaId));
+						super.LOG.log(Level.WARNING, MessageFormat.format("Unable to find existing temporary media for temporaryID: {0}", tempMediaId));
 					}
 				}
 			}
 			// Save new html to description
-			String newDescription = descriptionDoc.toString();
-			attachedComponent.setDescription(newDescription);
-			component.getComponent().setDescription(newDescription);
+			if (attachedComponent != null) {
+				String newDescription = descriptionDoc.toString();
+				attachedComponent.setDescription(newDescription);
+				component.getComponent().setDescription(newDescription);
 
-			// Delete temporary media that was converted.
-			for (String temporaryId : processedConversions.keySet()) {
-				componentService.getSystemService().removeTemporaryMedia(temporaryId);
+				// Delete temporary media that was converted.
+				for (String temporaryId : processedConversions.keySet()) {
+					componentService.getSystemService().removeTemporaryMedia(temporaryId);
+				}
+
+				persistenceService.persist(attachedComponent);
 			}
-
-			persistenceService.persist(attachedComponent);
 
 			if (approved) {
 				sendApprovalNotification(component.getComponent());
@@ -914,7 +919,7 @@ public class CoreComponentServiceImpl
 	public void cascadeDeleteOfComponent(String componentId, ComponentDeleteOptions option)
 	{
 		Objects.requireNonNull(componentId, "Component Id is required.");
-		log.log(Level.INFO, MessageFormat.format("Attempting to Removing component: {0}", componentId));
+		super.LOG.log(Level.INFO, MessageFormat.format("Attempting to Removing component: {0}", componentId));
 
 		Collection<Class<?>> entityClasses = DBManager.getConnection().getEntityManager().getRegisteredEntities();
 		for (Class entityClass : entityClasses) {
@@ -1207,7 +1212,7 @@ public class CoreComponentServiceImpl
 			try {
 				BeanUtils.setProperty(componentTrackingOrderExample, sortField.getName(), QueryByExample.getFlagForType(sortField.getType()));
 			} catch (IllegalAccessException | InvocationTargetException ex) {
-				log.log(Level.WARNING, "Unable to set sort field.", ex);
+				super.LOG.log(Level.WARNING, "Unable to set sort field.", ex);
 			}
 			queryByExample.setOrderBy(componentTrackingOrderExample);
 		}
@@ -1445,10 +1450,10 @@ public class CoreComponentServiceImpl
 								componentService.getUserService().checkComponentWatches(component);
 								componentsToIndex.add(component);
 							} else {
-								log.log(Level.FINE, MessageFormat.format("Component not found to update last Activity. Component may have been removed.  Check component id: {0}", componentId));
+								super.LOG.log(Level.FINE, MessageFormat.format("Component not found to update last Activity. Component may have been removed.  Check component id: {0}", componentId));
 							}
 						} catch (Exception e) {
-							log.log(Level.SEVERE, "Fail to update component.  Check data on component id: " + componentId, e);
+							super.LOG.log(Level.SEVERE, "Fail to update component.  Check data on component id: " + componentId, e);
 						}
 					}
 					componentService.getSearchService().indexComponents(componentsToIndex);
@@ -1760,7 +1765,7 @@ public class CoreComponentServiceImpl
 				}
 			} else {
 				String componentName = getComponentName(versionHistory.getComponentId());
-				log.log(Level.WARNING, MessageFormat.format("There is no files in the snapshot for component: {0} version: {1} ", componentName, versionHistory.getVersionHistoryId()));
+				super.LOG.log(Level.WARNING, MessageFormat.format("There is no files in the snapshot for component: {0} version: {1} ", componentName, versionHistory.getVersionHistoryId()));
 			}
 		}
 
@@ -1790,7 +1795,7 @@ public class CoreComponentServiceImpl
 				}
 			} else {
 				String componentName = getComponentName(versionHistory.getComponentId());
-				log.log(Level.WARNING, MessageFormat.format("There is no files in the snapshot for component: {0} version: {1} ", componentName, versionHistory.getVersionHistoryId()));
+				super.LOG.log(Level.WARNING, MessageFormat.format("There is no files in the snapshot for component: {0} version: {1} ", componentName, versionHistory.getVersionHistoryId()));
 			}		
 
 			if (archivedVersion != null) {
@@ -1835,7 +1840,7 @@ public class CoreComponentServiceImpl
 										source.toFile().cp_rp(destination);
 
 									} catch (IOException ex) {
-										log.log(Level.WARNING, MessageFormat.format("Failed to copy media to path file: {0}", mediaFile.getName()), ex);
+										super.LOG.log(Level.WARNING, MessageFormat.format("Failed to copy media to path file: {0}", mediaFile.getName()), ex);
 									}
 								}
 							}
@@ -1852,7 +1857,7 @@ public class CoreComponentServiceImpl
 										source.toFile().cp_rp(destination);
 
 									} catch (IOException ex) {
-										log.log(Level.WARNING, MessageFormat.format("Failed to copy resource to path file: {0}", resourceFile.getName()), ex);
+										super.LOG.log(Level.WARNING, MessageFormat.format("Failed to copy resource to path file: {0}", resourceFile.getName()), ex);
 									}
 								}
 							}
@@ -1860,7 +1865,7 @@ public class CoreComponentServiceImpl
 					}
 				} else {
 					String componentName = getComponentName(versionHistory.getComponentId());
-					log.log(Level.WARNING, MessageFormat.format("There is no files in the snapshot for component: {0} version: {1} ", componentName, versionHistory.getVersionHistoryId()));
+					super.LOG.log(Level.WARNING, MessageFormat.format("There is no files in the snapshot for component: {0} version: {1} ", componentName, versionHistory.getVersionHistoryId()));
 				}
 				
 				//save old version (keep in mind the update date will reflect now.)
@@ -1891,7 +1896,7 @@ public class CoreComponentServiceImpl
 			if (path != null) {
 				if (path.toFile().exists()) {
 					if (path.toFile().delete() == false) {
-						log.log(Level.WARNING, "Unable to delete component version: {0}", path.toString());
+						super.LOG.log(Level.WARNING, "Unable to delete component version: {0}", path.toString());
 					}
 				}
 			}
@@ -2005,7 +2010,7 @@ public class CoreComponentServiceImpl
 		for (int i=targetEntities.size()-1; i >=0; i--) {
 			if (StringUtils.isBlank(targetEntities.get(i).uniqueKey())) {
 				T badRecord = targetEntities.remove(i);
-				log.log(Level.WARNING, MessageFormat.format("Bad record (found during merge...it was removed): {0}", StringProcessor.printObject(badRecord)));				
+				super.LOG.log(Level.WARNING, MessageFormat.format("Bad record (found during merge...it was removed): {0}", StringProcessor.printObject(badRecord)));				
 			}
 		}		
 		
@@ -2077,14 +2082,81 @@ public class CoreComponentServiceImpl
 		return componentType;
 	}
 
-	public void removeComponentType(String componentType)
+	public void removeComponentType(String componentType, String newComponentType)
 	{
 		ComponentType componentTypeFound = persistenceService.findById(ComponentType.class, componentType);
 		if (componentTypeFound != null) {
-			componentTypeFound.setActiveStatus(ComponentType.INACTIVE_STATUS);
-			componentTypeFound.populateBaseUpdateFields();
-			persistenceService.persist(componentTypeFound);
+			
+			boolean inactivate = true;
+			if (StringUtils.isNotBlank(newComponentType)) {
+				ComponentType newType = persistenceService.findById(ComponentType.class, newComponentType);
+				
+				if (newType != null) {
+					//migrate data
+					Component setComponent = new Component();
+					setComponent.setComponentType(newComponentType);
+					
+					Component whereComponent = new Component();
+					whereComponent.setComponentType(componentType);
+					
+					persistenceService.updateByExample(Component.class, setComponent, whereComponent);
+														
+					
+					//remove restrictions
+					AttributeType attributeTypeExample = new AttributeType();
+					List<AttributeType> allAttributes = attributeTypeExample.findByExample();
+					List<AttributeType> updateAttributes = new ArrayList<>();
+					for (AttributeType attributeType : allAttributes) {
+						
+						boolean addToUpdate = false;
+						if (attributeType.getRequiredRestrictions() != null) {
+							for (int i=attributeType.getRequiredRestrictions().size()-1; i >= 0; i--) {
+								String checkType = attributeType.getRequiredRestrictions().get(i).getComponentType();
+								if (checkType.equals(componentType))
+								{
+									attributeType.getRequiredRestrictions().remove(i);
+									addToUpdate = true;
+								}
+							}
+						}
+						
+						if (attributeType.getAssociatedComponentTypes() != null) {
+							for (int i=attributeType.getAssociatedComponentTypes().size()-1; i >= 0; i--) {
+								String checkType = attributeType.getAssociatedComponentTypes().get(i).getComponentType();
+								if (checkType.equals(componentType))
+								{
+									attributeType.getAssociatedComponentTypes().remove(i);
+									addToUpdate = true;
+								}
+							}						
+						}
+						
+						if (addToUpdate) {
+							updateAttributes.add(attributeType);
+						}
+						
+						for (AttributeType attributeTypeUpdated : updateAttributes) {
+							componentService.getAttributeService().saveAttributeType(attributeTypeUpdated, false);							
+						}
+					}
 
+					//remove
+					inactivate=false;
+					persistenceService.delete(componentTypeFound);
+				} else {
+					LOG.log(Level.WARNING, "Unable to find new component type: " 
+							+ newComponentType 
+							+ "  to migrate data to.  Inactivating component type: " 
+							+ componentType);
+				}
+			} 
+			
+			if (inactivate) {			
+				componentTypeFound.setActiveStatus(ComponentType.INACTIVE_STATUS);
+				componentTypeFound.populateBaseUpdateFields();
+				persistenceService.persist(componentTypeFound);
+			}
+			OSFCacheManager.getComponentCache().removeAll();
 			OSFCacheManager.getComponentTypeCache().removeAll();
 		}
 	}
@@ -2188,7 +2260,7 @@ public class CoreComponentServiceImpl
 
 			Component mainComponent = persistenceService.findById(Component.class, pendingChangeComponent.getPendingChangeId());
 			if (mainComponent != null) {
-				log.log(Level.FINEST, "Updating Core Information");
+				super.LOG.log(Level.FINEST, "Updating Core Information");
 				mainComponent.setName(pendingChangeComponent.getName());
 				mainComponent.setDescription(pendingChangeComponent.getDescription());
 				mainComponent.setComponentType(pendingChangeComponent.getComponentType());
@@ -2204,19 +2276,19 @@ public class CoreComponentServiceImpl
 				fileHistoryOption.setSkipRequiredAttributes(Boolean.TRUE);
 				doSaveComponent(requiredForComponent, fileHistoryOption);
 
-				log.log(Level.FINEST, "Merge Component");
+				super.LOG.log(Level.FINEST, "Merge Component");
 				fileHistoryOption = new FileHistoryOption();
 				fileHistoryOption.setUploadTags(Boolean.TRUE);
 				fileHistoryOption.setSkipRequiredAttributes(Boolean.TRUE);
 				mergedComponent = merge(componentIdOfPendingChange, mainComponent.getComponentId(), fileHistoryOption);
 
-				log.log(Level.FINEST, "Send notification of approval (if requested)");
+				super.LOG.log(Level.FINEST, "Send notification of approval (if requested)");
 				//Temp change just to carry of the notification request
 				mergedComponent.setNotifyOfApprovalEmail(pendingChangeComponent.getNotifyOfApprovalEmail());
 				sendChangeRequestNotification(mergedComponent);
 
 			} else {
-				log.log(Level.WARNING, "Parent component doesn't exist unable to merge pending change. (Removing pending change)");
+				super.LOG.log(Level.WARNING, "Parent component doesn't exist unable to merge pending change. (Removing pending change)");
 				cascadeDeleteOfComponent(componentIdOfPendingChange);
 			}
 		} else {
